backend/migrations/models/1_20221122010144_update.py
from tortoise import BaseDBAsyncClient


async def upgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "fabcard" ADD "cardRarity" VARCHAR(64) NOT NULL;"""


async def downgrade(db: BaseDBAsyncClient) -> str:
    return """
        ALTER TABLE "fabcard" DROP COLUMN "cardRarity";"""
